# [SNOW-1678113] Improve pivot_table performance