--- a/src/snowflake/snowpark/modin/plugin/_internal/ordered_dataframe.py
+++ b/src/snowflake/snowpark/modin/plugin/_internal/ordered_dataframe.py
@@ -483,10 +483,6 @@ class OrderedDataFrame:
         """Returns a Snowpark session object associated with this ordered dataframe."""
         return self._dataframe_ref.snowpark_dataframe.session
 
-    @cached_property
-    def all_active_column_snowflake_quoted_identifiers(self) -> list[str]:
-        return self._get_active_column_snowflake_quoted_identifiers()
-
     def _get_active_column_snowflake_quoted_identifiers(
         self,
         include_ordering_columns: bool = True,
@@ -545,7 +541,7 @@ class OrderedDataFrame:
         return column_quoted_identifiers
 
     def _extract_quoted_identifiers_from_column_or_name(
-        self, col: ColumnOrName
+        self, col: ColumnOrName, active_columns: list[str]
     ) -> list[str]:
         """
         Extract the snowflake quoted identifiers out of a Column or column name with following rule:
@@ -577,7 +573,6 @@ class OrderedDataFrame:
                 # star adds all projected columns to the result
                 return self.projected_column_snowflake_quoted_identifiers
             else:
-                active_columns = self.all_active_column_snowflake_quoted_identifiers
                 if col in active_columns:
                     return [col]
                 else:
@@ -620,6 +615,7 @@ class OrderedDataFrame:
         new_snowpark_column_objects: list[Column] = []
         # a list of snowflake quoted identifiers as projected columns for new OrderedDataFrame
         new_projected_columns: list[str] = []
+        active_columns = self._get_active_column_snowflake_quoted_identifiers()
         for e in exprs:
             if isinstance(e, TableFunctionCall):
                 # we couldn't handle TableFunctionCall, so just use the original select
@@ -628,7 +624,9 @@ class OrderedDataFrame:
                 )
                 return OrderedDataFrame(DataFrameReference(snowpark_dataframe))
             elif isinstance(e, (Column, str)):
-                column_names = self._extract_quoted_identifiers_from_column_or_name(e)
+                column_names = self._extract_quoted_identifiers_from_column_or_name(
+                    e, active_columns
+                )
                 new_projected_columns.extend(column_names)
                 if isinstance(e, Column):
                     new_snowpark_column_objects.append(e)
@@ -733,9 +731,12 @@ class OrderedDataFrame:
         exprs = parse_positional_args_to_list(*agg_exprs)
         # extract the aggregation function name
         result_column_quoted_identifiers: list[str] = []
+        active_columns = self._get_active_column_snowflake_quoted_identifiers()
         for e in exprs:
             if isinstance(e, (Column, str)):
-                column_names = self._extract_quoted_identifiers_from_column_or_name(e)
+                column_names = self._extract_quoted_identifiers_from_column_or_name(
+                    e, active_columns
+                )
                 result_column_quoted_identifiers.extend(column_names)
             else:
                 raise AssertionError(
@@ -793,7 +794,7 @@ class OrderedDataFrame:
         # check all ordering_columns are in the current active columns
         _raise_if_identifier_not_exists(
             [column.snowflake_quoted_identifier for column in ordering_columns],
-            self.all_active_column_snowflake_quoted_identifiers,
+            self._get_active_column_snowflake_quoted_identifiers(),
             "ordering column",
         )