--- a/src/snowflake/snowpark/modin/plugin/_internal/utils.py
+++ b/src/snowflake/snowpark/modin/plugin/_internal/utils.py
@@ -1902,7 +1902,6 @@ def append_columns(
         column_objects
     ), f"The number of column identifiers ({len(column_identifiers)}) is not equal to the number of column objects ({len(column_objects)})"
 
-    # existing_columns = df.projected_column_snowflake_quoted_identifiers
     new_columns = [
         column_object.as_(column_identifier)
         for column_identifier, column_object in zip(column_identifiers, column_objects)