--- a/tests/integ/test_cte.py
+++ b/tests/integ/test_cte.py
@@ -22,9 +22,9 @@ from tests.integ.utils.sql_counter import SqlCounter, sql_count_checker
 from tests.utils import IS_IN_STORED_PROC, IS_IN_STORED_PROC_LOCALFS, TestFiles, Utils
 
 pytestmark = [
-    pytest.mark.skipif(
+    pytest.mark.xfail(
         "config.getoption('local_testing_mode', default=False)",
-        reason="CTE is a SQL feature and SQL counter check doesn't work for local testing mode",
+        reason="CTE is a SQL feature",
         run=False,
     ),
     pytest.mark.skipif(